Moving onSettled into a shared function.
