+++ sdk/servicebus/service-bus/src/core/batchingReceiver.ts
@@ -5,7 +5,6 @@ import * as log from "../log";
 import { MessagingError, translate } from "@azure/core-amqp";
 import {
   AmqpError,
-  Delivery,
   EventContext,
   OnAmqpEvent,
   ReceiverEvents,
@@ -16,7 +15,6 @@ import { ReceiveMode, ServiceBusMessageImpl } from "../serviceBusMessage";
 import {
   MessageReceiver,
   OnAmqpEventAsPromise,
-  PromiseLike,
   ReceiveOptions,
   ReceiverType
 } from "./messageReceiver";
@@ -24,6 +22,7 @@ import { ClientEntityContext } from "../clientEntityContext";
 import { throwErrorIfConnectionClosed } from "../util/errors";
 import { AbortSignalLike } from "@azure/core-http";
 import { checkAndRegisterWithAbortSignal } from "../util/utils";
+import { sharedOnSettled } from "./shared";
 
 /**
  * Describes the batching receiver where the user can receive a specified number of messages for
@@ -380,41 +379,7 @@ export class BatchingReceiver extends MessageReceiver {
       const onSettled: OnAmqpEvent = (context: EventContext) => {
         const connectionId = this._context.namespace.connectionId;
         const delivery = context.delivery;
-        if (delivery) {
-          const id = delivery.id;
-          const state = delivery.remote_state;
-          const settled = delivery.remote_settled;
-          log.receiver(
-            "[%s] Delivery with id %d, remote_settled: %s, remote_state: %o has been " +
-              "received.",
-            connectionId,
-            id,
-            settled,
-            state && state.error ? state.error : state
-          );
-          if (settled && this._deliveryDispositionMap.has(id)) {
-            const promise = this._deliveryDispositionMap.get(id) as PromiseLike;
-            clearTimeout(promise.timer);
-            log.receiver(
-              "[%s] Found the delivery with id %d in the map and cleared the timer.",
-              connectionId,
-              id
-            );
-            const deleteResult = this._deliveryDispositionMap.delete(id);
-            log.receiver(
-              "[%s] Successfully deleted the delivery with id %d from the map.",
-              connectionId,
-              id,
-              deleteResult
-            );
-            if (state && state.error && (state.error.condition || state.error.description)) {
-              const error = translate(state.error);
-              return promise.reject(error);
-            }
-
-            return promise.resolve();
-          }
-        }
+        sharedOnSettled(connectionId, delivery, this._deliveryDispositionMap);
       };
 
       const addCreditAndSetTimer = (reuse?: boolean): void => {
@@ -532,47 +497,4 @@ export function getRemainingWaitTimeInMsFn(
 
     return Math.min(remainingTimeMs, maxTimeAfterFirstMessageInMs);
   };
-}
-
-/**
- * @internal
- * @ignore
- */
-export function onSettled() {
-  const connectionId = this._context.namespace.connectionId;
-  const delivery = context.delivery;
-  if (delivery) {
-    const id = delivery.id;
-    const state = delivery.remote_state;
-    const settled = delivery.remote_settled;
-    log.receiver(
-      "[%s] Delivery with id %d, remote_settled: %s, remote_state: %o has been " + "received.",
-      connectionId,
-      id,
-      settled,
-      state && state.error ? state.error : state
-    );
-    if (settled && this._deliveryDispositionMap.has(id)) {
-      const promise = this._deliveryDispositionMap.get(id) as PromiseLike;
-      clearTimeout(promise.timer);
-      log.receiver(
-        "[%s] Found the delivery with id %d in the map and cleared the timer.",
-        connectionId,
-        id
-      );
-      const deleteResult = this._deliveryDispositionMap.delete(id);
-      log.receiver(
-        "[%s] Successfully deleted the delivery with id %d from the map.",
-        connectionId,
-        id,
-        deleteResult
-      );
-      if (state && state.error && (state.error.condition || state.error.description)) {
-        const error = translate(state.error);
-        return promise.reject(error);
-      }
-
-      return promise.resolve();
-    }
-  }
 }
