+++ sdk/servicebus/service-bus/src/core/messageReceiver.ts
@@ -30,6 +30,7 @@ import { MessageHandlerOptions } from "../models";
 import { DispositionStatusOptions } from "./managementClient";
 import { AbortSignalLike } from "@azure/core-http";
 import { AbortError } from "@azure/abort-controller";
+import { sharedOnSettled, PromiseLike } from "./shared";
 
 /**
  * @internal
@@ -48,15 +49,6 @@ interface CreateReceiverOptions {
  */
 export interface OnAmqpEventAsPromise extends OnAmqpEvent {
   (context: EventContext): Promise<void>;
-}
-
-/**
- * @internal
- */
-export interface PromiseLike {
-  resolve: (value?: any) => void;
-  reject: (reason?: any) => void;
-  timer: NodeJS.Timer;
 }
 
 /**
@@ -286,40 +278,8 @@ export class MessageReceiver extends LinkEntity {
     this._onSettled = (context: EventContext) => {
       const connectionId = this._context.namespace.connectionId;
       const delivery = context.delivery;
-      if (delivery) {
-        const id = delivery.id;
-        const state = delivery.remote_state;
-        const settled = delivery.remote_settled;
-        log.receiver(
-          "[%s] Delivery with id %d, remote_settled: %s, remote_state: %o has been " + "received.",
-          connectionId,
-          id,
-          settled,
-          state && state.error ? state.error : state
-        );
-        if (settled && this._deliveryDispositionMap.has(id)) {
-          const promise = this._deliveryDispositionMap.get(id) as PromiseLike;
-          clearTimeout(promise.timer);
-          log.receiver(
-            "[%s] Found the delivery with id %d in the map and cleared the timer.",
-            connectionId,
-            id
-          );
-          const deleteResult = this._deliveryDispositionMap.delete(id);
-          log.receiver(
-            "[%s] Successfully deleted the delivery with id %d from the map.",
-            connectionId,
-            id,
-            deleteResult
-          );
-          if (state && state.error && (state.error.condition || state.error.description)) {
-            const error = translate(state.error);
-            return promise.reject(error);
-          }
 
-          return promise.resolve();
-        }
-      }
+      sharedOnSettled(connectionId, delivery, this._deliveryDispositionMap);
     };
 
     this._onAmqpMessage = async (context: EventContext) => {
